+++ table/block_based_table_factory.cc
@@ -383,14 +383,6 @@ std::string BlockBasedTableFactory::GetPrintableTableOptions() const {
   snprintf(buffer, kBufferSize, "  block_align: %d\n",
            table_options_.block_align);
   ret.append(buffer);
-  snprintf(buffer, kBufferSize, "  memory_allocator: %p\n",
-           table_options_.memory_allocator.get());
-  ret.append(buffer);
-  if (table_options_.memory_allocator) {
-    snprintf(buffer, kBufferSize, "  memory_allocator_name: %s\n",
-             table_options_.memory_allocator->Name());
-    ret.append(buffer);
-  }
   return ret;
 }